+++ src/main/kotlin/io/hhplus/concert/domain/seat/SeatRepository.kt
@@ -4,9 +4,9 @@ import io.hhplus.concert.domain.schedule.Schedule
 
 interface SeatRepository {
     fun save(seat: Seat): Seat
-    fun save(schedule: Schedule, number: Int): Seat
+    fun save(schedule: Schedule, number: Long): Seat
 
-    fun findByScheduleIdAndNumber(scheduleId: Long, number: Int): Seat?
+    fun findByScheduleIdAndNumber(scheduleId: Long, number: Long): Seat?
 
     fun findAllByScheduleId(scheduleId: Long): List<Seat>
 }
